+++ CHANGELOG.md
@@ -10,6 +10,7 @@ For a list of breaking changes, check [here](#breaking-changes).
 - [#1883](https://github.com/clj-kondo/clj-kondo/issues/1883): constructor usage should have name-col in analysis
 - [#1874](https://github.com/clj-kondo/clj-kondo/issues/1874): fix name of fully qualified class usage
 - [#1876](https://github.com/clj-kondo/clj-kondo/issues/1876): suppress . analysis from .. macroexpansion
+- [#1877](https://github.com/clj-kondo/clj-kondo/issues/1876): suppress `new` analysis from `(String. x)` expansion
 
 ## 2022.11.02